Added support for a handshake function.
The function can be exported by an injection DLL and it will get called with information about the target executable and the hooks that are to be used. The DLL can than decide whether the target is appropriate and also provide a message that may contain version information or an error message.

The handshake function has to be defined as
extern "C" __declspec(dllexport) HRESULT __cdecl SyringeHandshake(SyringeHandshakeInfo* pInfo)

Valid return values are S_OK, if the DLL is to be used, or S_FALSE, if it is not to be used. Return E_* values to indicate errors (the DLL is not used in this case, and no Message is put in the log file).