Fixed a regression regarding thread handles
This was introduced with b5ad829f.

Threat handles in the threadInfoMap were closed manually when the SyringeDebugger instance was destroyed. By using the ThreadHandle, this is done automatically, even when handling the EXIT_THREAD_DEBUG_EVENT. It previously wasn't handled there. When calling ContinueDebugEvent after such an event, the handle was closed already.

By releasing ownership from the ThreadHandle, the handle is no longer closed.